docs(rename folder): rename middleware folder to plural
