+++ server/routes/user.router.ts
@@ -9,7 +9,7 @@
 import { Router } from 'express';
 // eslint-disable-next-line import/no-unresolved
 import userController from '../controllers/user.controller';
-import authMiddleware from '../middleware/auth';
+import authMiddleware from '../middlewares/auth';
 
 const userRouter = Router();
 // --------------------------------------------------------
